chore: 엔티티 필드 값 타입을 모두 래퍼타입으로 변경

+++ src/main/java/com/quickticket/quickticket/domain/review/entity/Review.java
@@ -25,7 +25,7 @@ public class Review {
     private User user;
 
     @Column(columnDefinition = "TINYINT UNSIGNED", nullable = false)
-    private byte userRating;
+    private Short userRating;
 
     @Column(columnDefinition = "TIMESTAMP", nullable = false)
     private LocalDateTime createdAt;
